track the renderer route in main via did-navigate-in-page
