+++ packages/app/index.ts
@@ -175,7 +175,7 @@ async function init() {
 
   if (!app.commandLine.hasSwitch("disable-bring-to-top-on-focus")) {
     main.window.on("focus", () => {
-      if (appState.visibleSurface === "account") {
+      if (main.visibleSurface === "account") {
         accounts.getSelectedAccount().instance.gmail.view.webContents.focus();
       }
     });
