+++ packages/app/menu.ts
@@ -22,7 +22,6 @@ import { openExternalUrl } from "@/url";
 import { WorkspaceApp } from "@/workspace-app";
 import { licenseKey } from "./license-key";
 import { createMeruMessageUrl } from "./protocol";
-import { appState } from "./state";
 
 export class AppMenu {
   private _menu: Menu | undefined;
@@ -165,7 +164,7 @@ export class AppMenu {
 
     const isGmailVisible =
       focusedWindow === main.window &&
-      appState.visibleSurface === "account" &&
+      main.visibleSurface === "account" &&
       selectedAccount.instance.tabs.activeTab.id === GMAIL_TAB_ID;
 
     const userEmail = selectedAccount.instance.gmail.userEmail;
