+++ packages/renderer/lib/ipc.ts
@@ -6,14 +6,6 @@ ipc.renderer.on("navigate", (_event, to) => {
   navigate(to);
 });
 
-function sendRouteChanged() {
-  ipc.main.send("route.changed", `/${window.location.hash.replace(/^#?\/?/, "")}`);
-}
-
-window.addEventListener("hashchange", sendRouteChanged);
-
-sendRouteChanged();
-
 ipc.renderer.on("taskbar.setOverlayIcon", (_event, unreadCount) => {
   const canvas = document.createElement("canvas");
